pyproject.toml
[tool.poetry]
name = "ibhack-mcp-server"
version = "1.0.0"
description = "A basic MCP server with test tools"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ibhack_mcp_server", from = "."}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastmcp = "^2.12.2"
google-generativeai = "^0.8.3"
aiohttp = "^3.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mcp-server = "server:mcp"
